backend/evo_worker/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
httpx
sqlalchemy
psycopg2-binary
numpy==1.26.4
pandas